requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
Babel==2.8.0
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
defusedxml==0.6.0
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1
django-admin-thumbnails==0.2.5
django-allauth==0.42.0
django-countries==6.1.2
django-heroku==0.3.1
django-likert-field==0.2.0
django-phone-field==1.8.1
django-phonenumber-field==4.0.0
django-taggit==1.3.0
filelock==3.0.12
gunicorn==20.0.4
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
phonenumbers==8.12.7
Pillow==7.2.0
psycopg2==2.8.5
pylint==2.5.3
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
stripe==2.50.0
toml==0.10.1
urllib3==1.25.10
virtualenv==20.0.31
whitenoise==5.2.0
wrapt==1.12.1
yapf==0.30.0
